Spring Boot #23 #25

+++ papaya/src/main/java/msp/papaya/repository/GPSRepository.java
@@ -3,5 +3,5 @@ package msp.papaya.repository;
 import msp.papaya.model.GPS;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GPSRepository extends JpaRepository<GPS, Long> {
+public interface GPSRepository extends JpaRepository<GPS, Integer> {
 }
